// src/types/responseEnums.ts
enum ResponseEnum {
    BAD_REQUEST = 'BAD_REQUEST',
    INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
    DATABASE_ERROR = 'DATABASE_ERROR',
    CACHE_ERROR = 'CACHE_ERROR',
    NOT_FOUND = 'NOT_FOUND',
    FORBIDDEN = 'FORBIDDEN',
    UNAUTHORIZED = 'UNAUTHORIZED',
    SUCCESS = 'SUCCESS',
    BANNED_TERRITORY = 'BANNED_TERRITORY',
    ALLOWED_TERRITORY = 'ALLOWED_TERRITORY',
    INVALID_COORDINATES = 'INVALID_COORDINATES',
    COORDINATES_MISSING = 'COORDINATES_MISSING'
}

export default ResponseEnum;